Fixes for unit and feature tests
