--- test/test_printers.py
@@ -1,1 +1,0 @@
-import auto_qc.printers
